+++ frontend_tests/node_tests/muting.js
@@ -9,7 +9,6 @@ const blueslip = require("../zjsunit/zblueslip");
 const {page_params} = require("../zjsunit/zpage_params");
 
 const user_topics = zrequire("user_topics");
-const muted_users = zrequire("muted_users");
 const stream_data = zrequire("stream_data");
 
 const design = {
@@ -45,7 +44,6 @@ stream_data.add_sub(social);
 function test(label, f) {
     run_test(label, ({override}) => {
         user_topics.set_user_topics([]);
-        muted_users.set_muted_users([]);
         f({override});
     });
 }
@@ -53,9 +51,6 @@ function test(label, f) {
 test("edge_cases", () => {
     // private messages
     assert.ok(!user_topics.is_topic_muted(undefined, undefined));
-
-    // invalid user
-    assert.ok(!muted_users.is_user_muted(undefined));
 });
 
 test("add_and_remove_mutes", () => {
@@ -77,45 +72,6 @@ test("add_and_remove_mutes", () => {
     // test unknown stream is harmless too
     user_topics.remove_muted_topic(unknown.stream_id, "java");
     assert.ok(!user_topics.is_topic_muted(unknown.stream_id, "java"));
-
-    assert.ok(!muted_users.is_user_muted(1));
-    muted_users.add_muted_user(1);
-    assert.ok(muted_users.is_user_muted(1));
-
-    // test idempotency
-    muted_users.add_muted_user(1);
-    assert.ok(muted_users.is_user_muted(1));
-
-    muted_users.remove_muted_user(1);
-    assert.ok(!muted_users.is_user_muted(1));
-
-    // test idempotency
-    muted_users.remove_muted_user(1);
-    assert.ok(!muted_users.is_user_muted(1));
-});
-
-test("get_unmuted_users", () => {
-    const hamlet = {
-        user_id: 1,
-        full_name: "King Hamlet",
-    };
-    const cordelia = {
-        user_id: 2,
-        full_name: "Cordelia, Lear's Daughter",
-    };
-    const othello = {
-        user_id: 3,
-        full_name: "Othello, Moor of Venice",
-    };
-
-    muted_users.add_muted_user(hamlet.user_id);
-    muted_users.add_muted_user(cordelia.user_id);
-
-    assert.deepEqual(
-        muted_users.filter_muted_user_ids([hamlet.user_id, cordelia.user_id, othello.user_id]),
-        [othello.user_id],
-    );
-    assert.deepEqual(muted_users.filter_muted_users([hamlet, cordelia, othello]), [othello]);
 });
 
 test("get_mutes", () => {
@@ -140,48 +96,6 @@ test("get_mutes", () => {
             stream: office.name,
             stream_id: office.stream_id,
             topic: "gossip",
-        },
-    ]);
-
-    assert.deepEqual(muted_users.get_muted_users(), []);
-    muted_users.add_muted_user(6, 1577836800);
-    muted_users.add_muted_user(4, 1577836800);
-    const all_muted_users = muted_users
-        .get_muted_users()
-        .sort((a, b) => a.date_muted - b.date_muted);
-
-    assert.deepEqual(all_muted_users, [
-        {
-            date_muted: 1577836800000,
-            date_muted_str: "Jan\u00A001,\u00A02020",
-            id: 6,
-        },
-        {
-            date_muted: 1577836800000,
-            date_muted_str: "Jan\u00A001,\u00A02020",
-            id: 4,
-        },
-    ]);
-});
-
-test("unknown streams", () => {
-    page_params.muted_users = [
-        {id: 3, timestamp: 1577836800},
-        {id: 2, timestamp: 1577836800},
-    ];
-
-    muted_users.initialize();
-
-    assert.deepEqual(muted_users.get_muted_users().sort(), [
-        {
-            date_muted: 1577836800000,
-            date_muted_str: "Jan\u00A001,\u00A02020",
-            id: 3,
-        },
-        {
-            date_muted: 1577836800000,
-            date_muted_str: "Jan\u00A001,\u00A02020",
-            id: 2,
         },
     ]);
 });
